Move events next to types part I
