+++ types/src/front/assistant/conversation.ts
@@ -272,3 +272,41 @@ export interface FetchConversationMessagesResponse {
   lastValue: number | null;
   messages: MessageWithRankType[];
 }
+
+/**
+ * Conversation events.
+ */
+
+// Event sent when the user message is created.
+export type UserMessageNewEvent = {
+  type: "user_message_new";
+  created: number;
+  messageId: string;
+  message: UserMessageWithRankType;
+};
+
+// Event sent when the user message is created.
+export type UserMessageErrorEvent = {
+  type: "user_message_error";
+  created: number;
+  error: {
+    code: string;
+    message: string;
+  };
+};
+
+// Event sent when a new message is created (empty) and the agent is about to be executed.
+export type AgentMessageNewEvent = {
+  type: "agent_message_new";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  message: AgentMessageWithRankType;
+};
+
+// Event sent when the conversation title is updated.
+export type ConversationTitleEvent = {
+  type: "conversation_title";
+  created: number;
+  title: string;
+};
