+++ types/src/index.ts
@@ -32,6 +32,7 @@ export * from "./front/assistant/agent";
 export * from "./front/assistant/avatar";
 export * from "./front/assistant/builder";
 export * from "./front/assistant/conversation";
+export * from "./front/assistant/pubsub";
 export * from "./front/assistant/templates";
 export * from "./front/assistant/visualization";
 export * from "./front/content_fragment";
@@ -51,10 +52,7 @@ export * from "./front/lib/api/assistant/actions/process";
 export * from "./front/lib/api/assistant/actions/retrieval";
 export * from "./front/lib/api/assistant/actions/tables_query";
 export * from "./front/lib/api/assistant/actions/websearch";
-export * from "./front/lib/api/assistant/agent";
-export * from "./front/lib/api/assistant/conversation";
 export * from "./front/lib/api/assistant/generation";
-export * from "./front/lib/api/assistant/pubsub";
 export * from "./front/lib/api/credentials";
 export * from "./front/lib/assistant";
 export * from "./front/lib/connectors_api";
